labhub: Change regex pattern for 'invite_cmd'
The original file used regex with '\s*' responsible for matching 0 or
more whitespaces and led to matches like 'invite @userto developers' as
there are 0 backspaces between '@user' and 'to'. With '\s+' atleast 1
whitespace is required for the regex to match such commands. Added a
test to verify.

Fixes https://github.com/coala/corobo/issues/512