tests: Isolate tests from loading labhub
IsolatedTestCase prevents TestBot from loading
LabHub as IGitt imports were delaying loading
time for the bot due to repetitive attempts to
send requests until an HTTP Client Error is
raised.

It was also not possible to patch the IGitt
imports while testing these plugins as plugins in
Errbot are loaded dynamically using yapsy plugin
manager. So it is not possible to patch since the
namespace is unknown.

IsolatedTestCase is able to prevent loading of
labhub plugin by renaming its `.plug` file. This
prevents the TestBot from loading it.